requirements.txt
numpy
scipy
matplotlib
scikit-learn
tensorflow-gpu
jupyter
